Implemented useFavorite hook
